Don't use different manifest file name for SI and SR component
Use same manifest filename and determine it's SI or SR by its
contents.
And delete utils. APIs that in utils should be in service impl file.
