+++ components/ntp_background_images/browser/BUILD.gn
@@ -10,8 +10,6 @@ static_library("browser") {
     "ntp_background_images_service.h",
     "ntp_background_images_source.cc",
     "ntp_background_images_source.h",
-    "ntp_background_images_utils.cc",
-    "ntp_background_images_utils.h",
     "sponsored_images_component_data.cc",
     "sponsored_images_component_data.h",
     "switches.cc",
