--- components/ntp_background_images/browser/ntp_background_images_utils.h
@@ -1,22 +1,0 @@
-// Copyright (c) 2020 The Brave Authors. All rights reserved.
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this file,
-// you can obtain one at http://mozilla.org/MPL/2.0/.
-
-#ifndef BRAVE_COMPONENTS_NTP_BACKGROUND_IMAGES_BROWSER_NTP_BACKGROUND_IMAGES_UTILS_H_
-#define BRAVE_COMPONENTS_NTP_BACKGROUND_IMAGES_BROWSER_NTP_BACKGROUND_IMAGES_UTILS_H_
-
-#include <memory>
-
-#include "base/values.h"
-
-class PrefRegistrySimple;
-
-namespace ntp_background_images {
-
-void RegisterLocalStatePrefs(PrefRegistrySimple* registry);
-bool IsValidSuperReferralComponentInfo(const base::Value& component_info);
-
-}  // namespace ntp_background_images
-
-#endif  // BRAVE_COMPONENTS_NTP_BACKGROUND_IMAGES_BROWSER_NTP_BACKGROUND_IMAGES_UTILS_H_
